apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
yt-dlp==2023.12.30
requests==2.31.0
aioredis==2.0.1
celery==5.3.4
redis==5.0.1
sendgrid==6.11.0
stripe==7.4.0
python-dotenv==1.0.0
email-validator==2.1.0
